refactor: trigger dedupe service
