+++ packages/server/api/src/app/workers/flow-worker/flow-queue-consumer.ts
@@ -27,7 +27,7 @@ import { redisQueueManager } from './queues/redis/redis-queue'
 import { QueueMode, SystemProp, enrichErrorContext, exceptionHandler, logger, system } from 'server-shared'
 import { flowService } from '../../flows/flow/flow.service'
 import { triggerHooks } from '../../flows/trigger'
-import { dedupeService } from '../../flows/trigger/trigger-deduple.service'
+import { dedupeService } from '../../flows/trigger/dedupe'
 
 const queueMode = system.getOrThrow<QueueMode>(SystemProp.QUEUE_MODE)
 
@@ -208,7 +208,6 @@ const consumePieceTrigger = async (data: RepeatingJobData): Promise<void> => {
     const filterPayloads = await dedupeService.filterUniquePayloads(
         data.flowVersionId,
         payloads,
-        RunEnvironment.PRODUCTION,
     )
     const createFlowRuns = filterPayloads.map((payload) =>
         flowRunService.start({
